requirements.txt
requests
schedule
python-dotenv==0.19.2
PyGithub==1.59.1
openai==1.3.0
